I53 - TP4

+++ I53-Theorie_du_language_et_compilation/TP/TP4/lib/arrayUtils.h
@@ -1,13 +1,14 @@
 #ifndef ARRAY_UTILS_H
 #define ARRAY_UTILS_H
 
+#include <stdio.h>
 #include "utils.h"
 
 bool intIsInArray(int value, int *array, size_t arraySize);
 
 void fillIntArray(int *array, size_t arraySize, int value);
 
-void shiftValuesArrayToRight(int *array, size_t arraySize, size_t startShiftingIndex);
+void shiftValuesArrayToRight(int *array, int arraySize, int startShiftingIndex);
 
 void printArray(int *array, size_t arraySize);
